requirements.txt
flask==2.2.3
werkzeug==2.2.3
flask-sqlalchemy==3.0.3
numpy==1.24.3
pandas==2.0.0
openpyxl==3.1.2
PyPDF2==3.0.1
pymysql==1.0.3
psycopg2-binary==2.9.6
gunicorn==20.1.0